+++ src/csharp/core/TickTrader.Algo.Backtester/Entities/Dealer/OpenOrderDealerRequest.cs
@@ -16,7 +16,7 @@ namespace TickTrader.Algo.Backtester
         public Quote CurrentRate { get; }
 
         public bool Confirmed { get; private set; }
-        public decimal? DealerAmount { get; private set; }
+        public double? DealerAmount { get; private set; }
         public double? DealerPrice { get; private set; }
 
         public void Confirm()
@@ -24,7 +24,7 @@ namespace TickTrader.Algo.Backtester
             Confirmed = true;
         }
 
-        public void Confirm(decimal amount, double price)
+        public void Confirm(double amount, double price)
         {
             DealerAmount = amount;
             DealerPrice = price;
